+++ instapy/instapy.py
@@ -4270,6 +4270,7 @@ class InstaPy:
         # Get `followers` data
         grabbed_followers = get_followers(
             self.browser,
+            self.username,
             username,
             amount,
             self.relationship_data,
@@ -4318,6 +4319,7 @@ class InstaPy:
         # get `following` data
         grabbed_following = get_following(
             self.browser,
+            self.username,
             username,
             amount,
             self.relationship_data,
@@ -4349,6 +4351,7 @@ class InstaPy:
         # get all and active Unfollowers
         all_unfollowers, active_unfollowers = get_unfollowers(
             self.browser,
+            self.username,
             username,
             compare_by,
             compare_track,
@@ -4373,6 +4376,7 @@ class InstaPy:
         # get Nonfollowers
         nonfollowers = get_nonfollowers(
             self.browser,
+            self.username,
             username,
             self.relationship_data,
             live_match,
@@ -4397,6 +4401,7 @@ class InstaPy:
         # get Fans
         fans = get_fans(
             self.browser,
+            self.username,
             username,
             self.relationship_data,
             live_match,
@@ -4421,6 +4426,7 @@ class InstaPy:
         # get Mutual Following
         mutual_following = get_mutual_following(
             self.browser,
+            self.username,
             username,
             self.relationship_data,
             live_match,
